+++ src/cluster/Cluster.h
@@ -42,8 +42,8 @@ public:
     bool operator==(Cluster &other);
     Cluster& operator=(Cluster &other);
 
-    static float distance(Cluster &c1, Cluster &c2, AbstractMetric *pMetric);
-    float diameter(AbstractMetric *pMetric);
+    static double distance(Cluster &c1, Cluster &c2, AbstractMetric *pMetric);
+    double diameter(AbstractMetric *pMetric);
 
 private:
     void init();
