+++ src/cluster/validity.h
@@ -8,18 +8,18 @@ struct ThreadDistanceData {
     public:
         ThreadDistanceData(Cluster *, Cluster *, AbstractMetric *);
         static void* threaded_distance(void*);
-        float distance();
+        double distance();
         Cluster* pCluster1;
         Cluster* pCluster2;
         AbstractMetric *pMetric;
     private:
-        float _distance;
+        double _distance;
 };
 
 class Validity {
     public:
-        static float dunn(Clustering &, AbstractMetric *);
-        static float bezderk(Clustering &, AbstractMetric *);
+        static double dunn(Clustering &, AbstractMetric *);
+        static double bezderk(Clustering &, AbstractMetric *);
 
 };
